Use FParsec
Learnt about FParsec and decided to use that instead of hand writing the
parser. It wasn't without some trepidation, because I worried that it
will cause problems when Bash throws a curve ball in it's grammar, but
the GNU Bash project uses Yacc to generate a parser, so it should be
possible to use FParsec (a parser generator) instead of a hand written parser.